+++ Changelog.md
@@ -18,6 +18,7 @@ v2.37.4 (not published yet)
 ====================
 
 - UI: Trigger event "While Processing" more often, but also time limited
+- UI: Make timeout for error messages on job processing optional
 
 
 v2.37.3 (2024-03-23)